Commit: Added episode_advantages method to PPO class.
The method episode_advantages computes advantages using generalised
advantage estimation given the rewards of shape (n) and the critic
values of shape (n+1) corresponding to each step of the env in an
episode. Automatically uses the cpu device even if cuda is available
since advantage estimation is a backwards recursion calculation.

Added a first test for the episode_advantages method and a test for
making sure cuda is available.