Commit: Added a train method to PPO class.
The training process so far seems to produce expected results for both
the gymnasium CartPole-v1 and Acrobot-v1 environments.

Todo:

Should try to use pytorch parallel mapping function when computing
the advantages for each terminated/truncated episode.

Should only be sending the gathered rollout data to cuda device the
moment it is required. Numpy should be used as an intermediary data
storage.